requirements.txt
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
google-generativeai==0.8.6
